requirements.txt
Flask==3.0.2
spotipy==2.23.0
Flask-Cors==4.0.0
python-dotenv==1.0.1